Cleaning up code, adding delete options

+++ Internship-7-Library/InternshipLibrary.Domain/Repositories/AuthorRepository.cs
@@ -14,39 +14,43 @@ namespace InternshipLibrary.Domain.Repositories
     {
         private readonly LibraryContext _context;
 
-        public AuthorRepository()
-        {
-            _context = new LibraryContext();
-        }
+        public AuthorRepository() => _context = new LibraryContext();
 
-        public void Create(Author author)
+        public string Create(Author author)
         {
+            if (author.FirstName == "" || author.LastName == "")
+                return "You must enter a first and last name";
+            if (Read().FirstOrDefault(x => x.FirstName == author.FirstName && x.LastName == author.LastName) != null)
+                return "That author already exists";
             _context.Authors.Add(author);
             _context.SaveChanges();
+            return "Successfully added";
         }
 
-        public Author Read(int id)
-        {
-            return _context.Authors.Find(id);
-        }
+        public Author Read(int id) => _context.Authors.Find(id);
 
-        public List<Author> Read()
-        {
-            return _context.Authors.ToList();
-        }
+        public List<Author> Read() => _context.Authors.ToList();
 
-        public void Update(Author newAuthor, Author oldAuthor)
+        public string Update(Author newAuthor, Author oldAuthor)
         {
-            newAuthor.Id = oldAuthor.Id;
-            _context.Remove(oldAuthor);
-            _context.Authors.Add(newAuthor);
+            if (newAuthor.FirstName == "" || newAuthor.LastName == "")
+                return "You must enter a first and last name";
+            if (Read().FirstOrDefault(x => x.FirstName == newAuthor.FirstName && x.LastName == newAuthor.LastName) != null)
+                return "You didn't change anything";
+
+            oldAuthor.FirstName = newAuthor.FirstName;
+            oldAuthor.LastName = newAuthor.LastName;
             _context.SaveChanges();
+            return "Successfully updated";
         }
 
-        public void Delete(Author author)
+        public string Delete(Author author)
         {
+            if (_context.Books.Any(x => x.Author == author))
+                return "You must first delete all the authors books";
             _context.Authors.Remove(author);
             _context.SaveChanges();
+            return "Successfully removed";
         }
     }
 }
