+++ Internship-7-Library/InternshipLibrary.Domain/Repositories/PublisherRepository.cs
@@ -13,39 +13,42 @@ namespace InternshipLibrary.Domain.Repositories
     {
         private readonly LibraryContext _context;
 
-        public PublisherRepository()
-        {
-            _context = new LibraryContext();
-        }
+        public PublisherRepository() => _context = new LibraryContext();
 
-        public void Create(Publisher publisher)
+        public string Create(Publisher publisher)
         {
+            if (publisher.Name == "")
+                return "Please enter a name";
+            if (Read().FirstOrDefault(x => x.Name == publisher.Name) != null)
+                return "A publisher with the same name already exists, enter a unique name";
             _context.Publishers.Add(publisher);
             _context.SaveChanges();
+            return "Successfully added";
         }
 
-        public Publisher Read(int id)
-        {
-            return _context.Publishers.Find(id);
-        }
+        public Publisher Read(int id) => _context.Publishers.Find(id);
 
-        public List<Publisher> Read()
-        {
-            return _context.Publishers.ToList();
-        }
+        public List<Publisher> Read() => _context.Publishers.ToList();
 
-        public void Update(Publisher newPublisher, Publisher oldPublisher)
+        public string Update(Publisher newPublisher, Publisher oldPublisher)
         {
-            newPublisher.Id = oldPublisher.Id;
-            _context.Publishers.Remove(oldPublisher);
-            _context.Publishers.Add(newPublisher);
+            if (newPublisher.Name == "")
+                return "Please enter a name";
+            if (Read().FirstOrDefault(x => x.Name == newPublisher.Name) != null)
+                return "You haven't made any changes";
+
+            oldPublisher.Name = newPublisher.Name;
             _context.SaveChanges();
+            return "Successfully updated";
         }
 
-        public void Delete(Publisher publisher)
+        public string Delete(Publisher publisher)
         {
+            if (_context.Books.Any(x => x.Publisher == publisher))
+                return "You must first delete all of the books which have that publisher";
             _context.Publishers.Remove(publisher);
             _context.SaveChanges();
+            return "Successfully removed";
         }
     }
 }
